xxsaa

+++ Shiping company Data Structure/Node.h
@@ -4,18 +4,18 @@ class Node
 {
 	T data;
 	Node<T>* next;
-	int weight;
+	double weight;
 public:
 	Node();
 	Node(const T& item);
 	Node(const T& item,Node<T>* nxt);
-	Node(const T& item, int w);
+	Node(const T& item, double w);
 	Node<T>* getnext() const;
 	T getitem() const;
 	void setnext(Node<T>* ptr);
 	void setitem(const T& item);
-	void SetWeight(int w);
-	int GetWeight();
+	void SetWeight(double w);
+	double GetWeight();
 };
 
 template<typename T>
@@ -37,7 +37,7 @@ template<typename T>
 }
 
  template<typename T>
- Node<T>::Node(const T& item, int w)
+ Node<T>::Node(const T& item, double w)
  {
 	 data = item;
 	 weight = w;
@@ -67,13 +67,14 @@ template<typename T>
 		data = item;
   }
   template<typename T>
-  void Node<T>::SetWeight(int w)
+  void Node<T>::SetWeight(double w)
   {
 	  weight = w;
   }
+ 
 
   template<typename T>
-  int Node<T>::GetWeight()
+  double Node<T>::GetWeight()
   {
 	  return weight;
   }
