+++ Shiping company Data Structure/PriorityQueueADT.h
@@ -5,7 +5,7 @@ class PriorityQueueADT
 public:
 	virtual bool peek(T & item) const = 0;
 	virtual bool dequeue(T& item) = 0;
-	virtual bool enqueue(const T & item,int weight) = 0;
+	virtual bool enqueue(const T & item,double weight) = 0;
 	virtual bool isempty() = 0;
 	virtual int GetSize()=0;
 };
